Updated api validation

+++ NwbaApi/Models/DataManager/CustomerManager.cs
@@ -22,8 +22,10 @@ namespace NwbaApi.Models.DataManager
             return _context.Customers.Find(id);
         }
 
+        // rreturns address for the customer ID
         public Address GetAddress(int id)
         {
+
             var customer = _context.Customers.Include(x => x.Address).FirstOrDefault(x => x.CustomerID == id);
             customer.Address.Customers = null; // To eliminate possible object cycle error
             return customer.Address;
@@ -41,21 +43,39 @@ namespace NwbaApi.Models.DataManager
 
             return customer.CustomerID;
         }
-
+        // This method find the customer, if exists updates the customer
+        // else return -1 for failure 
         public int Update(int id, Customer customer)
         {
-            _context.Update(customer);
-            _context.SaveChanges();
-
-            return id;
+            var c = Get(id);
+            if(c != null)
+            {
+                _context.Update(customer);
+                _context.SaveChanges();
+                return id;
+            }
+            else
+            {
+                return -1;
+            }
+            
         }
 
+        // This method find the customer, if exists delete the customer and returns customer id 
+        // else return -1 for failure 
         public int Delete(int id) 
         {
-            _context.Customers.Remove(_context.Customers.Find(id));
-            _context.SaveChanges();
-
-            return id;
+            var customer = Get(id);
+            if(customer != null)
+            {
+                _context.Customers.Remove(_context.Customers.Find(id));
+                _context.SaveChanges();
+                return id;
+            }
+            else
+            {
+                return -1;
+            }
         }
